Smooth Movement Added
Oculus technical issues at the university have been taken care of.  This update has added smooth movement controls into the project so the "player" can move around in the environment using the left joystick on the oculus controls.  Also added snap camera turning to the right hand joystick. To speed development Unitys' XR Toolkit was used to implement this so the learning was focussed on how to use this system as it already.