Interactable object and hands added
An initial box to interact with has been added to the scene. This can be picked up though this isn't in an ideal location currently. This will be moved in future to offset position to "float" away from the hand with a visual effect that makes sense for the games appearance(magic causing the float). Hands were added in for the visuals and the positioning resolved to match up with where VR controllers are. Animations created for the hands for when "grabbing" an item as well as a pointer for setting up interactions such as pressing keys on a keypad.