Initial UI Created
The initial UI for the spell creation screen has been created.  This has been populated with simple buttons for testing with a text output for showing which spell code is created from the "rune slots" these will be updated further down the line with sprites showing runes and spells.  The step on from this will be testing spells on the left hand. Also another animation for the hand has been created for when the hand has been idle for a while.