Keypad system added and animation tested
The keypad system was added into the game tested alongside the hand animations. The hand now goes into a point and the player can tap numbers with the  extended finger to type in the numbers. This also tested the animations on the hand to go in and out of this.  If the user enters an incorrect number the keypad clears, if correct it removes an in game barrier.