Spells and UI Update
Spells tested to work correctly with VR system, some issues with positioning have been correct. The UI has been updated to use correct rune symbols and spell names.  This is to be replicated onto hands if possible, enemy AI to be added from separate testing file on home pc now this is integrated correctly and final spells to be integrated as well. Once this is done final level map to be placed and tested.